Enhancements that help make things "just work" for me with EBS-mounted master database
Explicitly wait for AWS to complete the requested mount.  This might be faster or slower than 10 seconds, but it's certainly more correct.

Runs mkfs.xfs and use its output to tell if it's already got an fs on there.  This works in case the user created the volume in an external tool, with the volume-id already pasted into deploy.rb

Replaced the symlink method that didn't work, with a copying method that works.  It does, of course, put a damper on the prettiness of being able to directly edit the current sudoers file.  Maybe a hard link would be better, but this does function as-is.